Add tests for finding paths with multiple relationships
Between the same entities.
